Revert "HOTT-726: fill changes table"
