+++ app/models/change_log.rb
@@ -5,7 +5,7 @@ class ChangeLog
 
   def initialize(changes = [])
     @changes = changes.map { |change_attributes|
-      ChangeOld.new(change_attributes)
+      Change.new(change_attributes)
     }
   end
