--- db/migrate/20210610150945_create_changes.rb
@@ -1,13 +1,0 @@
-Sequel.migration do
-  change do
-    create_table :changes do
-      primary_key :id
-      String  :goods_nomenclature_item_id
-      Integer :goods_nomenclature_sid
-      String  :productline_suffix, size: 255
-      Bool    :end_line
-      String  :change_type,        size: 255
-      Date    :change_date
-    end
-  end
-end
