+++ db/structure.sql
@@ -837,35 +837,6 @@ ALTER SEQUENCE public.certificates_oid_seq OWNED BY public.certificates_oplog.oi
 
 
 --
--- Name: changes; Type: TABLE; Schema: public; Owner: -
---
-
-CREATE TABLE public.changes (
-    id integer NOT NULL,
-    goods_nomenclature_item_id text,
-    goods_nomenclature_sid integer,
-    productline_suffix character varying(255),
-    end_line boolean,
-    change_type character varying(255),
-    change_date date
-);
-
-
---
--- Name: changes_id_seq; Type: SEQUENCE; Schema: public; Owner: -
---
-
-ALTER TABLE public.changes ALTER COLUMN id ADD GENERATED BY DEFAULT AS IDENTITY (
-    SEQUENCE NAME public.changes_id_seq
-    START WITH 1
-    INCREMENT BY 1
-    NO MINVALUE
-    NO MAXVALUE
-    CACHE 1
-);
-
-
---
 -- Name: chapter_notes; Type: TABLE; Schema: public; Owner: -
 --
 
@@ -7711,22 +7682,6 @@ ALTER TABLE ONLY public.certificates_oplog
 
 
 --
--- Name: changes changes_pkey; Type: CONSTRAINT; Schema: public; Owner: -
---
-
-ALTER TABLE ONLY public.changes
-    ADD CONSTRAINT changes_pkey PRIMARY KEY (id);
-
-
---
--- Name: changes changes_upsert_unique; Type: CONSTRAINT; Schema: public; Owner: -
---
-
-ALTER TABLE ONLY public.changes
-    ADD CONSTRAINT changes_upsert_unique UNIQUE (goods_nomenclature_sid, change_date);
-
-
---
 -- Name: chapter_notes chapter_notes_pkey; Type: CONSTRAINT; Schema: public; Owner: -
 --
 
@@ -10778,5 +10733,3 @@ INSERT INTO "schema_migrations" ("filename") VALUES ('20210108165416_add_hjid_to
 INSERT INTO "schema_migrations" ("filename") VALUES ('20210108170325_add_hjid_to_geographical_area_memberships_view.rb');
 INSERT INTO "schema_migrations" ("filename") VALUES ('20210112135615_add_new_fields_to_geographical_area_memberships_oplog.rb');
 INSERT INTO "schema_migrations" ("filename") VALUES ('20210112160504_add_fields_to_geographical_area_memberships_view.rb');
-INSERT INTO "schema_migrations" ("filename") VALUES ('20210610150945_create_changes.rb');
-INSERT INTO "schema_migrations" ("filename") VALUES ('20210628165555_add_unique_constraint_to_changes.rb');
