--- lib/changes_table_populator/commodity_code_description_changed.rb
@@ -1,25 +1,0 @@
-module ChangesTablePopulator
-  class CommodityCodeDescriptionChanged < Importer
-    class << self
-      def source_table
-        :goods_nomenclature_description_periods
-      end
-
-      def select_condition
-        -> { [goods_nomenclature_item_id, goods_nomenclature_sid, productline_suffix] }
-      end
-
-      def where_condition(day: Date.current)
-        { validity_start_date: day }
-      end
-
-      def import_records(elements:, day: Date.current)
-        elements.map { |element| integrate_element(row: element, day: day) }
-      end
-
-      def change_type
-        'commodity'
-      end
-    end
-  end
-end
