+++ lib/goods_nomenclature_mapper.rb
@@ -6,7 +6,7 @@ class GoodsNomenclatureMapper
     def children
       @children ||= []
     end
-
+    
     def child_ids
       children.map(&:goods_nomenclature_sid)
     end
@@ -70,7 +70,7 @@ class GoodsNomenclatureMapper
   end
 
   def map_goods_nomenclatures(primary, secondary)
-    if (heading_map?(primary.reload, secondary.reload) &&
+    if (heading_map?(primary, secondary) &&
        (primary.producline_suffix < secondary.producline_suffix)) ||
        (primary.number_indents < secondary.number_indents)
