+++ spec/factories/goods_nomenclature_factory.rb
@@ -88,43 +88,6 @@ FactoryBot.define do
     trait :with_heading do
       after(:create) do |commodity, _evaluator|
         FactoryBot.create(:heading, goods_nomenclature_item_id: "#{commodity.goods_nomenclature_item_id.first(4)}000000")
-        commodity.reload
-      end
-    end
-
-    trait :with_children do
-      after(:create) do |commodity, _evaluator|
-        # Prepare some intermediate item ids
-        item_id = commodity.goods_nomenclature_item_id.to_i
-
-        # Make the commodity a parent
-        commodity.producline_suffix = '80'
-        commodity.save
-        FactoryBot.create(:goods_nomenclature_indent,
-                          goods_nomenclature_sid: commodity.goods_nomenclature_sid,
-                          goods_nomenclature_item_id: commodity.goods_nomenclature_item_id,
-                          validity_start_date: commodity.validity_start_date,
-                          validity_end_date: commodity.validity_end_date,
-                          productline_suffix: commodity.producline_suffix,
-                          number_indents: 1)
-
-        # Add another intermediate level
-        FactoryBot.create(:commodity,
-                          :with_indent,
-                          goods_nomenclature_item_id: "#{item_id + 1}",
-                          producline_suffix: '10',
-                          indents: 2)
-
-        # Add two leaf commodities
-        FactoryBot.create(:commodity,
-                          :with_indent,
-                          goods_nomenclature_item_id: "#{item_id + 1}",
-                          indents: 3)
-        FactoryBot.create(:commodity,
-                          :with_indent,
-                          goods_nomenclature_item_id: "#{item_id + 2}",
-                          indents: 3)
-        commodity.reload
       end
     end
   end
